+++ package.json
@@ -4,7 +4,7 @@
   "repository": "rognstadragnar/classnames-in-js",
   "license": "MIT",
   "version": "0.1.3",
-  "source": "src/index.js",
+  "source": "src/react.js",
   "main": "dist/classnames-in-js.js",
   "module": "dist/classnames-in-js.esm.js",
   "devDependencies": {
@@ -30,28 +30,43 @@
   },
   "scripts": {
     "release": "np",
-    "bundle": "npm-run-all bundle:*",
-    "bundle:react": "microbundle",
-    "bundle:preact-cjs": "microbundle src/preact.js -o preact.js -f cjs --compress",
-    "bundle:preact-es": "microbundle src/preact.js -o preact.js -f es --compress",
-    "dev": "microbundle watch --compress false",
+    "prebuild": "npm run clean",
+    "build": "npm-run-all build:*",
+    "build:react": "microbundle",
+    "build:react-lite": "microbundle src/react-lite.js -o lite.js -f cjs --compress && microbundle src/react-lite.js -o lite.js -f es --compress",
+    "build:preact": "microbundle src/preact.js -o preact.js -f cjs --compress && microbundle src/preact.js -o preact.js -f es --compress",
+    "build:preact-lite": "microbundle src/preact-lite.js -o preact-lite.js -f cjs --compress && microbundle src/preact-lite.js -o preact-lite.js -f es --compress",
+    "verify": "npm-run-all verify:*",
     "verify:lint": "eslint ./src",
     "verify:test": "mocha -r @babel/register __tests__/**/*.spec.js",
-    "test": "npm run verify:lint && npm run verify:test",
+    "test": "npm run verify",
     "prepare": "npm run build",
-    "build": "npm run clean && npm run bundle",
     "clean": "npm-run-all clean:*",
     "clean:react": "rimraf dist",
-    "clean:preact": "rimraf dist preact.js preact.js.map preact.esm.js preact.esm.js.map preact.umd.js preact.umd.js.map"
+    "clean:react-lite": "rimraf dist lite.js lite.js.map lite.esm.js lite.esm.js.map lite.umd.js lite.umd.js.map",
+    "clean:preact": "rimraf dist preact.js preact.js.map preact.esm.js preact.esm.js.map preact.umd.js preact.umd.js.map",
+    "clean:preact-lite": "rimraf dist preact-lite.js preact-lite.js.map preact-lite.esm.js preact-lite.esm.js.map preact-lite.umd.js preact-lite.umd.js.map"
   },
   "files": [
     "src",
     "dist",
+    "lite.js",
+    "lite.js.map",
+    "lite.esm.js",
+    "lite.esm.js.map",
+    "lite.umd.js",
+    "lite.umd.js.map",
     "preact.js",
     "preact.js.map",
     "preact.esm.js",
     "preact.esm.js.map",
     "preact.umd.js",
-    "preact.umd.js.map"
+    "preact.umd.js.map",
+    "preact-lite.js",
+    "preact-lite.js.map",
+    "preact-lite.esm.js",
+    "preact-lite.esm.js.map",
+    "preact-lite.umd.js",
+    "preact-lite.umd.js.map"
   ]
 }
