--- src/index.js
@@ -1,7 +1,0 @@
-import styled from './tag'
-import { Theme } from './theme'
-
-export { styled, Theme }
-
-// type classNameFn<T>= (props: T) => string
-// type ClassName<T> = classNameFn<T> | string
