+++ src/lib/react-theme-context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+const Theme = createContext(null)
+
+export default Theme
