+++ src/react.js
@@ -1,9 +1,11 @@
 import React from 'react'
-import { getClassName, isPropValid } from './lib/index'
-import { Theme } from './theme'
+import tags from './lib/tags'
+import { isPropValid, cleanClassesTag } from './lib/index'
+import createRender from './lib/create-react-render'
+import Theme from './lib/react-theme-context'
 
-export default function styled(tag, styledClassName) {
-  const render = createRender(tag, styledClassName)
+function styledFn(tag, styledClassName) {
+  const render = createRender(tag, styledClassName, isPropValid)
   return React.forwardRef
     ? React.forwardRef(render)
     : // React.forwardRef won't available on older React versions and in Preact
@@ -11,33 +13,14 @@ export default function styled(tag, styledClassName) {
       ({ innerRef, ...rest }) => render(rest, innerRef)
 }
 
-function createRender(tag, styledClassName) {
-  function render(props, ref) {
-    const { as: component = tag, ...rest } = props
-    let filteredProps
+let styled = styledFn
 
-    // Check if it's an HTML tag and not a custom element
-    if (typeof tag === 'string' && !tag.includes('-')) {
-      filteredProps = {}
-
-      // eslint-disable-next-line guard-for-in
-      for (const key in rest) {
-        if (key === 'as' || isPropValid(key)) {
-          // Don't pass through invalid attributes to HTML elements
-          filteredProps[key] = rest[key]
-        }
-      }
-    } else {
-      filteredProps = rest
-    }
-
-    filteredProps.ref = ref
-    return React.createElement(Theme.Consumer, {
-      children: theme => {
-        filteredProps.className = getClassName(props, styledClassName, theme)
-        return React.createElement(component, filteredProps)
-      }
-    })
+tags.forEach(tag => {
+  styled[tag] = (...rest) => {
+    return styledFn(tag, (props, theme) =>
+      cleanClassesTag(rest, { props, theme })
+    )
   }
-  return render
-}
+})
+
+export { styled, Theme }
