--- src/theme.js
@@ -1,2 +1,0 @@
-import { createContext } from 'react'
-export const Theme = createContext(null)
